src/main/webapp/app/entities/autor/delete/autor-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { IAutor } from '../autor.model';
import { AutorService } from '../service/autor.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './autor-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class AutorDeleteDialogComponent {
  autor?: IAutor;

  constructor(protected autorService: AutorService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.autorService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
